Make global for class name have normal original
Summary: I discovered this while implementing the new version of global registration.  This evidently dates back to the original implementation of registering classes as meta types: D6643297.  I'm not sure why this was chosen, since it allows people to redefine their classes locally, which we don't support in a consistent manner.

Reviewed By: grievejia

Differential Revision: D17619648

fbshipit-source-id: 7a0be0b90902a9e1097e5589fe919f2099595a39