filtered callables don't break fixpoint invariant
Summary:
Our fixpoint invariant ensures that each callable that needs to be reanalyzed is actually within the set of all callables, to help detect bugs where a callable appears in the callgraph, but is missing from the set of callables to analyze.

This is all good and well, except in the case where a non-testing file calls a testing file, where we explicitly filter the callable.

Take the strategy of also computing which callables we filtered, and not fail the fixpoint if the missing callable is in the explicitly filtered set.

Reviewed By: fahndrich

Differential Revision: D17635444

fbshipit-source-id: 9b0b00b3a38a631284c9b07738d00450d21c3abc