Add missing dependencies
Summary:
We forgot to add these to the pile of final dependencies.
The fact that this is so easy to forget highlights how much we need a dependency searching module.  I also think that it isn't actually worth it to maintain separate types for each kind of dependency for each environment.

Reviewed By: sinancepel

Differential Revision: D17672957

fbshipit-source-id: 2db8c127d396248ecfb452a20d0277a87c033c3c