Unifying all of the dependency keys
Summary:
Now that we've finished up all of the environments, I think it's become relatively clear that maintaining separate types for each of the different environment's downstream dependencies is worth the pain.  We never take advantage of the fact that we've encoded that a environment may only have dependencies from downstream environments, and the one-way ness of the environment building is better reified by the typing of the actual `update` functions.

On the other hand, maintaining separate types has serious downsides:
* perf: each type maintains a separate encode/decode table, which means that the same dependency could have an equivalent entry in each previous table, which is not ideal.
* conversions (perf/ergonomics): we need to do conversions between the dependency types when encoding transitivity :(
* ergonomics: Having separate types makes extracting all previous dependencies confusing and error prone.  I haven't taken full advantage of the new system in this diff to keep it narrowly focused on a replacement, but a follow up diff will make dependency collection much much easier.

Reviewed By: sinancepel

Differential Revision: D17674375

fbshipit-source-id: 58d57aa522d8622cd96f21cf969200dacf418acb