Split module ignore codes from mode
Summary:
We'd like to enforce a single mode: strict, unsafe, or declare, on each module.

Setting module-wide codes to be ignored is independent of these three modes and shouldn't be conflated; you can only pick one mode but you should be allowed to set both `strict` and `ignore-all[15]`, for example.

We can also rip this out entirely when it is time.

Reviewed By: dkgi

Differential Revision: D17681945

fbshipit-source-id: 9e37ca78750bd9d427e39ff3d9121c5d101dd246