Abstracting the common interface of (most) environments
Summary:
There is currently a fair bit of repetition between the environments.  Now that we're trying to move this towards a maintainable solution, it's worth thinking about how to DRY this out.  This is I think a fairly uncontroversial abbreviation of the interfaces, and will hopefully become a good piece of summary documentation for the interfaces of these interfaces in general.  I think it'll be a bigger trade off when I start trying to functorize the actual implementations.

Note that I'm ignoring ASTEnvironment and UnannotatedGlobalEnvironment for now in order to make the threading of the fan-out easier to express.  I may revisit this.

Reviewed By: grievejia

Differential Revision: D17688400

fbshipit-source-id: b39e8ad866aa8c6cad4d201b7de54ea9ae968c3e