imprecisely handle key taint for dictionaries
Summary: This is a first approximation of how to handle key taint for dictionaries - simply join the precise, indexed taint with the key taint (as we don't represent keys well at the moment). If we find this to be an issue due to taint flowing everywhere from keys, we can consider adding a special index for key taint in dictionaries, and special-casing the `.keys()` function to return this taint.

Reviewed By: fahndrich

Differential Revision: D17703536

fbshipit-source-id: 8d89e90f6ac21991c5a5da00dc2bc9dbba484c79