Raise error when accessing private attributes of base classes.
Summary:
Problem: Accessing a "private" attribute of a base class did not raise a type error even though it caused an error at runtime (AttributeError: 'Child' object has no attribute '_Child__private').

This diff checks for a private attribute only within the current class, not in the base classes.

It also raises an error if the private attribute is accessed outside the base class in which the private attribute was defined. Otherwise, it is visible outside the class.

Reviewed By: sinancepel

Differential Revision: D17691582

fbshipit-source-id: 9172631af6edc4475e487913f84293a79faf4452