Keep track of extraneous local modes to error against
Summary:
Goal is to error when someone tries to use more than one pyre mode in a module. This has already started happening in distillery, when a codemod was applying `pyre-strict` to modules without checking if they were already annotated with `pyre-unsafe`. We've had a few independent requests to make this throw a pyre error.

The behavior this specifies is: take the mode you see first as the truth, error against any additional mode settings that follow.

RFC on whether this can be made less gross.

I couldn't think of a better way to set up errors on non-AST affecting parts of the codebase. We can't do it in type checking because the AST doesn't include these extraneous mode settings. It doesn't seem very clean to be re-iterating through the raw source of the entire module during the type checking of say, the toplevel define, either during type checking or post processing. The AST building layer also shouldn't be / can't be throwing any analysis errors.

Really, the source should be parsed once into an AST and any additional information we will need in the future packaged up into the metadata. Following that model, all we can do is add a gross additional field to hold these until we can error on them later on.

Also RFC on the name.

Reviewed By: MaggieMoss

Differential Revision: D17731850

fbshipit-source-id: f3978bca5c3ee25bb7870a5039a721ec3f70dc99