Make debug a consistent mode
Summary:
We allow using `# pyre-debug` on a module to get completely unfiltered errors (for debugging purposes, ofc, which is why this mode isn't publicized) on just that module, which makes sense because running a whole project with the `debug` flag can get unwieldy.

It doesn't make sense to have this separated out as a separate toplevel metadata field, and modify the configuration according to the individual module's setting. After all, a module can't be both strict and debug, so let's merge this in to be handled just like the other modes.

Reviewed By: sinancepel

Differential Revision: D17731228

fbshipit-source-id: 00e1407b46a7afc65636e3a73cb781e0d7fe1d0b