Eliminating backedges
Summary:
Backedges are problematic for a number of reasons:
* they're unparallelizable
* they're unlazifyable
* they're undependencytrackable

Even so, we kept them around because they allow us to efficiently implement the meeting of primitives.
However, it turns out that even for this purpose, doing a meet on two primitives that aren't less or equal one or the other is rarely used.
At this point, it seems like they have outlived their welcome.

Reviewed By: dkgi

Differential Revision: D17744878

fbshipit-source-id: 69bb8d78e00a722d088b2e4f10ed9c2a22f4046d