Don't warn if attribute was inherited from concrete parent
Summary: We were gathering attributes that are undefined from parents if they are abstract or protocols (to avoid double-erroring on things). Turns out we were using the same set of superclasses to look for initialized attributes to remove... this logic in general could be simplified with the ~transitive flag anyway.

Reviewed By: grievejia

Differential Revision: D17753108

fbshipit-source-id: fe88556fc044c3eff56b63774db09353d29a24f8