Kill default local mode using optional
Summary:
Realized while beginning to implement errors against redundant modes that having a 'Default' explicit mode gets really messy. We need to track locations on the other modes, and eventually this results in having to match away the variant to get at the location and throw lots of "failwiths" down the pipeline whenever we see a "Default" mode, which should be impossible to error against.

Make this a bit more true to form and just represent explicit modes with the variant, and keep it optional.

Reviewed By: sinancepel

Differential Revision: D17744327

fbshipit-source-id: 7905b87455f2d07edbd6f02807d6ae3d0402b46e