Eliminating global enabled flag for Type.Cache
Summary:
Having a global flag for the enabled-ness of the type cache is problematic and error prone.  This diff switches it over to be a flag on the individual `Type.create` call.

While I was revamping this, I realized that the current implementation wasn't even correctly clearing the cache.  Because there's a separate cache for each worker, calling clear in the master is not enough, you need to clear it in each of the workers.  Therefore when transitioning from explicitly clearing in each worker in AliasEnvironment (which may not even work if only a few aliases are being updated), I added a new iter to wipe the cache in each worker.

All of this should be fixed by making a common cache in shared memory that correctly propagates dependencies (the type cache swallows dependencies which is also bad :( )

Reviewed By: sinancepel

Differential Revision: D17765928

fbshipit-source-id: aa6e83e019fac3775eeaf9e2ff90dcad76d8fe70