Factor out table addition from register function
Summary:
Now that AliasEnvironment no longer has to juggle the Type.Cache, all of the register methods are now just iterating over the triggers, producing values and then adding those values to shared memory.

We can factor that behavior up a level into the Environment.Updater, which reduces duplication and allows us to move towards hiding the Table itself from each of the Enviroments.  This is also the interface we would need for lazification.

Reviewed By: dkgi

Differential Revision: D17766678

fbshipit-source-id: 962ef188149acc2e2d2da8480408f249420c05d0