Throw errors on extraneous local modes
Summary:
This should cap off the fixes/refactors related to better supporting local modes for Pyre.

We now error whenever a second mode is inadvertently added, pointing to the local mode we are using instead and where it was defined. We can clearly point to the configuration as the default, if no local mode exists. Error ignore codes are a separate system.

Reviewed By: dkgi

Differential Revision: D17744326

fbshipit-source-id: 0df290bbc329ee22bde9834ce8f62e5a39d8fff2