Remove -declare binary configuration flag
Summary:
We have made local modes an enum-like setting where you can only set one of
strict, unsafe, or declare. Sinan made a valid point that we should also be
refusing to validate any configurations that set more than one of these, as the
configuration setting is the default when no local mode is specified.

Investigated this, turns out 'strict: true' is the only configuration mode you
can set in a config. The absence of this is unsafe. I think that binary is
pretty clear. The other config options are debug and infer, which override how
the typechecker is running and ignores any local modes anyway.

The last case, declare, didn't actually make sense to me. Turns out it's not
even possible for the client to pass this in, you'd have to give it to the
OCaml binary on the command line. Typechecking your whole project under declare
by default seems to make no sense. This looks residual from when decl mode was
added to begin with and was never used since: D5885979

Reviewed By: grievejia

Differential Revision: D17789673

fbshipit-source-id: b91cec4f90e36c47edc68a23a93f9d85d9fc3a82