Re-Implementing consistency check functions for all environments via functor
Summary:
In the shuffle of moving all of our environments out, I neglected to maintain the ability to decode, show equality and provide a hash to key map.
Now that we have a functor that builds our update function, that's an easy location to implement this functionality.
I was able to recover all of the old functionality from version control and factor out the commonality into the functor.

That also constitutes the last thing that the raw table was being used for, so at the same time, I moved the table itself into the functor.  This required making two variants of the functor (one that uses a cached table and one that does not).

I still need to make the decoding functionality available to clients of the environments except for AnnotatedGlobalEnvironment, but that can wait until I functorize ReadOnly.

Reviewed By: dkgi

Differential Revision: D17768491

fbshipit-source-id: c258f7c0b7ddfff81bed733cdebfc8ae06fc6ef2