Standardize and ignore location of pyre logging directory
Summary:
note: this diff has been edited to just handle skipping the .pyre directory when type checking. Logic determining where that directory is is in the diff below

Spent some time thinking about how best to clean up .pyre directories spawned. This can help us cleanly kill servers and support multiple servers in fbcode. There are a couple scenarios to consider:

-  pyre configuration and local configuration set
- pyre configuration with sources set, and no local configs set (plus make sure the .pyre directory doesn't get typechecked)
- no configurations are set at all.

Currently we spin up .pyre directories in the local config folder, the project config folder, or when no configs are specified, the place pyre was called and the directory we call pyre on.

RFC: A better model for this would be to place all logs at the most nested configuration (edit: will use a single global .pyre with a nested directory hierarchy) - so local if it exists, project if it doesn't. This is convoluted, and we could reduce confusion in the future around this by reducing the distinction between a pyre config and a pyre local config - really the only rule we care about is that source directories are never overridden by a nested config. (This property also ensures that there is never a case where you have local configs and are also sometimes running against a project config).

When pyre is run with no configs specified at all (usually open-source), we should be dumping any necessary temporary files in /tmp (or just generate none at all) and not creating any .pyre directories elsewhere. This use-case is already just for supporting one-off trials of pyre, and it's reasonable to require a configuration if we want any stateful uses of pyre like spinning up a server. We already print a warning saying "try pyre init to create a config" - we can default to run a non-incremental check and clarify this warning.

General plan:
1. Add .pyre to ignore_all_errors
2. Centralize all references to .pyre to point to the constant .pyre directory set per configuration
3. Pass this directory into the binary so binary logging goes here too
4. If no configs exist, warn that we only support 'pyre check' without a server, and run that by default. Don't generate logs at all if possible (think about pyre infer...?)
5. (Make a bootcamp task to maybe check if a .pyre exists before overwriting it based on some key. Not high priority at all)

Reviewed By: dkgi

Differential Revision: D17795130

fbshipit-source-id: 7198c13803ce8e70e28995853b6057f52c2c0b2a