Using exposed decoding functions from environments
Summary: As promised in D17797436, we now can use the exposed decoding functionality.  This hypothetically could be done in the functor, but that will be trickier than I believe it will be worth because of the ClassHierarchyEnvironment case of a two table environment.  Because each one is only used once, I don't think it'll be that useful.

Reviewed By: sinancepel

Differential Revision: D17798589

fbshipit-source-id: eb7d0d99f1348c1ab255794eda8f482231ff2e1e