No longer filter out UninitializedAttribute s that are typed as Any
Summary: UninitializedAttribute errors contain a `mismatch` field, but it is not actually used in a "mismatching" way, as expected always equals actual.  However, the presence of the field led to confusion and its inclusion in a variant that would filter it if the expected/actual passed when Any was set to Bottom.  In effect this would allow any type containing Any to be have its uninitialized error filtered out.

Reviewed By: grievejia

Differential Revision: D17866423

fbshipit-source-id: 283f07b2a2deb038e9861bd3206070e7e3cf6bb9