Fixing Type.create cache corruption issue
Summary:
A cache only works if it is keyed on all inputs to the function it is caching.  Type.create is based on being passed a `aliases` function, which originally was only ever a lookup into the Aliases table.  However, I started hijacking it when  implementing generic parameter enforcement.  This meant that we were sometimes calling it with an argument which the cache was not aware of, leading to corruption issues where the cache was sometimes returning types with corrected aliases, and sometimes was not, unrelated to what we were actually asking for.

This was masking another issue, where we should have been always correcting the aliases in type checking.

My solution was to pull up the cache for this alias corrected version up into GlobalResolution, correctly separating the caches, and actually buying us some perf by avoiding re-doing the validation work that happened after the type create.

Reviewed By: grievejia

Differential Revision: D17889522

fbshipit-source-id: 3f24eb328741f814cb125d1a65d69cb529cbea8e