killing AnnotatedClass.Method
Summary: As far as I can tell, AnnotatedClass.Method is a hold-over from a time when it fulfilled the role of Attributes.  This is no longer necessary, as it is only used in a query that is not being actively used, and a taint analysis use case that was easily inlined.

Reviewed By: grievejia

Differential Revision: D17920705

fbshipit-source-id: 8c096a88e4b91aef05b5d1f04a7f3c169fb7ca91