Moving AnnotatedCallable.is_generator into the parser
Summary:
I'm trying to remove the need to store as many define bodies as we do, as those account for most user updates.
There is only one downstream consumer of the bodies of attribute defines, AnnotatedCallable.is_generator.  We can move this all the way up to the parser, since this is a purely syntactic property, and is as essential to interpreting a signature as whether a function is async.

Reviewed By: grievejia

Differential Revision: D17938995

fbshipit-source-id: 7996bdcb4c2c573064c71e50a9f54b0c88f73e5e