Only store signatures in Ast.Attributes, not defines
Summary:
Now that we have moved AnnotatedCallable.is_generator back to the parser in D17938995, we can swap out the defines currently stored in the Attribute with just the signatures.
This will be a memory win right now, and will enable decreased triggers when we start storing Attributes instead of class statements.

This required fixing up the downstream consumers of the attributes.

Reviewed By: grievejia

Differential Revision: D17944398

fbshipit-source-id: 1e3cbfc475fea14193df52c8db0366ae4f2d9767