No longer store full free function bodies in the UnannotatedGlobals table
Summary: This has the same purpose as D17973482, but is several orders of magnitude easier.  There are no fancy things that have to ever look at the bodies of free functions, so we can easily replace the Define.t with a Define.Signature.t, and (mostly) no-one blinks an eye.

Reviewed By: dkgi

Differential Revision: D17987428

fbshipit-source-id: d5d59c206d70eee6b7fbd521b72970a987a19e05