No longer store full class bodies in the ClassDefinitions table
Summary:
I am hypothesizing that a large proportion of updates are of bodies of functions.  This means that we ought to avoid storing these bodies if at all necessary, since our dependency tracking will pick up those changes as triggering downstream work, even if it is actually irrelevant.

The goal of this stack was to do this for a first target, class definitions.  It turned out that type checking only ever uses the body of the class for attribute building, so if we are able to push that upstream we can avoid triggering unnecessary type checking.

The trick here is that (already) attribute building is done in two stages:
* Ast.Statement.Class.t -> Ast.Statement.Attribute.t:   This is a purely syntactic translation of a class body into its component attributes.  This includes conventional attributes, implied attributes from constructors, defines, and properties.
* Ast.Statement.Attribute.t -> Annotated.Attribute.t:  This is a type-order-aware operation which consolidates the attributes of a class and its parents/metaclasses, and instantiates type variables, and more.

Currently we do both stages of this in type checking, but there is no reason why we can do the first stage in environment building (or even parsing 0_0).

This is exactly what this diff does :).

The complication comes from the fact that the first stage isn't actually all one stage.  The downstream user (AnnotatedClass), expects to be able to get just the explicit attributes, the implicit attributes, the explicit and implicit attributes, and the explicit and implicit attributes, if we've determined that this class inherits from a unittest class.

This means we can't just store one attribute map, we must instead store several components which we can combine on demand, which you can now see in Ast.Statement.Class.AttributeComponents.t.

These attribute components can be stored in a new type, ClassSummary.t.

By storing these attribute components instead of the bases, we ensure that an edit to the body of a method doesn't trigger anything beyond the UnannotatedGlobalEnvironment (unless of course it editing a constructor, and except type checking that directly edited file).

The goal of all of this can be seen in the following test:
```
  assert_updates
    ~original_source:
      {|
      class Foo:
        def method(self) -> None:
         print("hello")
    |}
    ~new_source:
      {|
      class Foo:
        def method(self) -> None:
         print("goodbye")
    |}
    ~middle_actions:[`Get ("test.Foo", dependency, Some 1)]
    ~expected_triggers:[]
    ~post_actions:[`Get ("test.Foo", dependency, Some 1)]
    ();
```
As you can see, updating the body of method does not trigger the dependency :)

Reviewed By: dkgi

Differential Revision: D17973482

fbshipit-source-id: 4cd6ef655b83803d677648dcb62d1f4be1012273