Don't return globals when global_fallback is false
Summary:
Looks like we still return globals sometimes when `global_fallback=false`.

This is causing us to be inconsistent about how we refine globals and attributes.

Really, this should have been avoided but the way we structure globals is making a local copy when we start analyzing the nested define (nested in the toplevel in most cases), which gets returned from `get_local ~global_fallback=false`.

Reviewed By: dkgi

Differential Revision: D17990354

fbshipit-source-id: d21664b98e21495ae9a6369b5e343d78e5e3911f