Add type variables for every state instead of nested state only
Summary:
Previously, when seeing a generic function like this:
```
T = TypeVar('T')
def foo(x: T) -> T: ...
```
We bound the type variable `T` when we **schedule** type checking of function `foo`. In other words, if `foo()` is defined in the toplevel of a class, we rely on the fact that `$toplevel` finished type checking first before `foo()` can gets checked, because scheduling `foo()` only happens when the `$toplevel` check is done.

Such a scheme does not work well if we want to decouple the type checking of `$toplevel` and `foo` -- the current behavior is that if `$toplevel` does not schedule the `foo()` check, variable `T` won't be registered for `foo()`, which would cause spurious errors when analyzing the `foo()` signature.

Fix this issue by moving the type variable registration logic from `State.forward_statement` to `State.initial`, which is called for every function regardless of their nesting status.

Reviewed By: mrkmndz

Differential Revision: D17992973

fbshipit-source-id: d05d3c4082192e3250ae2795086f1e41133c7605