Calculate log directory if missing when initializing logging
Summary:
logging can be set up from the shim, where we don't initialize arguments yet. In general it makes sense for the logging initializer to determine the log directory destination.

Sadly, it makes sense that determining where to log depends on filesystem operations (even if we did a bunch of refactoring, we'd still need to use `find_root` or duplicate it).

On the other hand, filesystem depends on log because it wants to log some performance metrics and that performance constant is a log global.

Did not see a fantastic way to get around this circular dependency, added a constants file to help with this instead. Open to alternative suggestions!

Reviewed By: dkgi

Differential Revision: D18009756

fbshipit-source-id: 5ffa6653f135776b9d5550bde327e685c7223dea