Restrict mode declarations to toplevel
Summary:
Ran into a bunch of un-ignoreable issues when upgrading distillery where some codemod script is setting a `# pyre-strict` in a big multi-line quote string.

I think the easiest way to resolve this is to mandate that `# pyre-X` mode settings are declared in toplevel. Can give some flexibility with a space prefix or not, but don't think it makes much sense to be declaring modes heavily indented or inside logic blocks since it applies to the whole file anyway.

Note: we were also lstripping the string before analyzing it for ignores/modes, which was throwing off all the locations before, too.

Reviewed By: grievejia

Differential Revision: D18141655

fbshipit-source-id: 43ec64219c58c52d2f6e175d17b136723143d41e