Register unannotated globals for try...catch statement
Summary:
We currently register unannotated globals for bodies of `if..else` statement, but not for bodies of `try...except` statements.

It is debatable whether we really want to look into those bodies. But IMO our decisions should at least be consistent: either we do it for both `if...else` and `try...except`, or we do neither. This diff chooses the pragmatic approach of looking into bodies of `try...except` as well, as it seems to have less impact than stop looking into `if...else` bodies.

Reviewed By: dkgi

Differential Revision: D18147412

fbshipit-source-id: d6a4a0fd5412c63c5e8b35f497b1d3e2c56fde35