Fix global annotation refinement
Summary:
I am surprised to learn in D17990354 that the way we structure globals is to make a local copy in `$toplevel`, and since we prefer local annotations over global annotations in `Resolution.get_local`, any local refinement that happens in `$toplevel` will take effect on those global variables, which is obviously not what we want. In addition, since local copy won't be typed before we finish type checking `$toplevel`, this behavior creates unnecessary dependency between type checking `$toplevel` and its descendants.

This diff fixes the issue by making sure we never refer to the `$toplevel` local copy in non-`$toplevel` functions: if we are not in `$toplevel` and find that a variable is global (via `GlobalResolution.global`), simply refuse to refine the type annotation.

Reviewed By: dkgi

Differential Revision: D18131629

fbshipit-source-id: 40ddd55b493a633f8bfed0d2ff3ea41211d5d3bc