Partially revert D17999114
Summary:
It turns out picking the last assignment to register as globals is a bad idea. Consider this:
```
x: Optional[int] = None
x = 1
```

The user annotates `x` as `Optional[int]`, but we pick up the definition of `x=1` and treat `x` as `int`, which is highly non-intuitive.

Although it's not hard to construct adversarial examples for the "picking the first assignment" behavior, I found it at least lead to fewer troubles than the `picking the last assignment" behavior.

Reviewed By: shannonzhu

Differential Revision: D18130487

fbshipit-source-id: d721cc54bd9889e8d324e9586d0c0dea43f40655