change Preprocessing.classes to preserve order of class definitions
Summary:
The core issue here is rough. We have some Python modules, such as setuptools, that conditionally re-define a class as an empty class based on a hasattr() call. For the taint analysis, we need to associate definitions with the actual class that defines them, which is not always possible due to the dynamic nature of Python.

This was the offending file (note the redefinition at the very end): https://github.com/pypa/setuptools/blob/master/setuptools/command/py36compat.py

Band-aid this issue by preferring the earlier definition over a later one, which fixes the setuptools crash. If we hit similar issues, we will probably need to consider all the classes and search for an implementation/store more metadata for callables.

Reviewed By: grievejia

Differential Revision: D18204209

fbshipit-source-id: a8637b274ea49b2bb0f5d026d5cae1eb47768b7e