Add location sensitive compare function to AST nodes
Summary:
Why do we need location sensitive compare/hash?
=======================================

Because locations attached to the AST nodes affect error messages. We want to make sure that when location changes, we update the locations in the errors accordingly. For example, let's say that our original file looks like this:

```
  def foo() -> None:
     ...

  def bar() -> int:
     return "foo"  #  <- Line X
```

There's going to be an "return type mismatch" error on line `X`. Now, if we edit this file like this:

```
  def foo() -> None:
+  # An empty line gets added here
     ...

  def bar() -> int:
     return "foo"  #  <- This becomes line X+1
```

Although we do not edit function `bar()` at all, the error message needs to be updated to point to line `X+1` instead. This is not possible today, because we perform location-insensitive comparison all the time -- which means that from our backend's perspective, the edit does not change anything. Therefore we'd refuse recomputing the errors.

Now, consider an actual AST change:

```
  def foo() -> None:
+  x = 42  # An assignment gets added here
     ...

  def bar() -> int:
     return "foo"  #  <- This becomes line X+1
```

We'll correctly adjust the line number for the error today, because we'd recompute type errors for the entire module regardless of which function gets the AST change. Notice that once we switch to function-level type checking, we won't be correct anymore: the AST of `bar()` didn't change in our edit so we are unlikely to re-process it. Therefore although location-sensitive hash/compare is not super important today, they will become more important in the course of pursuing fine-grained dependencies.

This diff
=======

This diff introduces `location_sensitive_compare` function to all relevant AST datastructures, i.e. `Node.t`, `Expression.t`, and `Statement.t`. The original `compare` functions ignore locations. The new `location_sensitive_compare` functions do not.

We have to manually write all the boilerplate code ourselves for `location_sensitive_compare` because the generated ones always invoke the location insensitive `compare` on their children -- what we want instead is to use the location sensitive version all the way down to each leaf node in the AST.

Reviewed By: sinancepel

Differential Revision: D18109283

fbshipit-source-id: 160d4e387cfce0325b15a73a7bdcd9cf18387f11