Report only one error about an Any in a NamedTuple attribute.
Summary:
Problem: When an initialized attribute in a NamedTuple uses the type Any, it throws a "Missing attribute annotation" error for the attribute and a "Missing parameter annotation" for the parameter of the implicit `__new__` method.

Solution: Don't throw the missing parameter error for `__new__` methods in a subclass of NamedTuple.

We identify that is a `__new__` method on a NamedTuple because the return type will be NamedTuple, as constructed in preprocessing.ml.

Reviewed By: dkgi

Differential Revision: D18235649

fbshipit-source-id: 922a1e3b9ec41545893cfc15b2a14b76fe78116c