support reporting paths by repository
Summary: Refactors how we instantiate locations for taint analysis JSONs by parametrizing the path lookup function. The basic change is adding a repository root flag and adding an alternate filename lookup function that gets the absolute path and relativizes to the repository root that the user specifies. Will revert to the existing behavior if no repo root is provided. The rest is plumbing.

Reviewed By: grievejia

Differential Revision: D18211572

fbshipit-source-id: b95cbdb52fde2bc01cd2b86f532166415f064a70