introduce ability to filter analysis by rule codes
Summary:
Since the taint analysis gets more expensive linearly in the number of sources and sinks, adding more rules and more exploratory sources & sinks was slowing the whole analysis down. Add a capability to filter the sources and sinks to the ones pertinent to a few rules a user is interested in.

I'm not attached to how I chose to implement the rule filtering - it felt easier to take care of it on the parsing side as subtracting from abstract map domains gets weird, but if people feel strongly, can change it to be a post-process filter.

Reviewed By: fahndrich

Differential Revision: D18231268

fbshipit-source-id: 68213786beae22ea1e388e2169b38cda69e65959