ensure stderr is flushed in case of failure
Summary:
Looks like we were not flushing stderr after the main process was finished and explicitly ignoring stderr output. This is probably OK in the success case, but when a subprocess has a non-zero exit, I suspect the user wants to know why.

Change the behavior so that if the returncode is nonzero, we join the stderr thread (therefore consistently printing the exception messages).

Reviewed By: grievejia

Differential Revision: D18234601

fbshipit-source-id: 8d8226396ecde6d85b6c9d9040987f7938dc0b58