Resolve exports with fewest changes possible
Summary:
Opt for using a module that exists rather than resolving an export if both are available.

I *think* we still need to keep the fixpoint widening threshold in for resolving exports on references that are incorrect. Whereas before, legitimate references could fall into this infinite loop, I think the reference has to be undefined now for this to happen (but we should still handle it). Didn't think too hard about this one, lmk if I'm wrong :P

Reviewed By: sinancepel

Differential Revision: D18226961

fbshipit-source-id: 1d9e708c44b8082588fdcb195d780e5238b541c9