Aggregate additional checks into one single check before invoking `Service.Check.run_check`
Summary:
We used to grab individual checks as a list in `Analysis.Check.checks`, run them all on a source, and concat the results.

I'm trying to change it such that `Analysis.Check.checks` (renamed to `create_check`) returns a single check instead of a list of check. This single check has a `run` function that does the same thing as if we run all checks then concat the results.

The motivation for this change is to prepare for diffs up the stack. More specifically, I'd like to separate post processing from individual checks. Aggregating all checks into one beforehand makes it easy to post-process once after all checks are done, instead of run post-process after each individual check.

Reviewed By: dkgi

Differential Revision: D18238857

fbshipit-source-id: 04ae31185e0dd5fe00315ac9727c9718903764ce