special case `x.__next__()` calls
Summary:
The `for element in container` pattern is syntactic sugar for `element = container.__iter__().__next()`. We were not properly simulating the fact that this is an arbitrary index access in the taint analysis.

Special case the last part of this pattern, the `__next__()` call, as it seems like the natural place to do so.

Reviewed By: dkgi

Differential Revision: D18252383

fbshipit-source-id: 8cf98b68afe7d13ecede02ee065aa60ca18e7bcd