Rationalizing the binary flags in AnnotatedAttribute.t
Summary:
Just as in Statement.Attribute.t, Annotated.Attribute.t had many combinations of fields that were either redundant or impossible.

There was a final field inside of `annotation` and one outside in Attribute, which was different from the "finalness" of `property = Some ReadOnly`.

To rectify this, I realized that there were really three different things we're trying to keep track of:
*  the "visibility" of setting capability, which has a nested relationship to whether a field can be refined (actually done in follow up diff), or  overridden in a subclass.
* whether an attribute came from an property (for taint analysis purposes)
* the actual resolved annotation types, both in "current" and "original" positions.

By making these explicit, I was able to iron out a bunch of bugs where we weren't erroring on missing attributes.

Reviewed By: dkgi

Differential Revision: D18261571

fbshipit-source-id: 47431d74dc2d1132fafee84b2028fcfc0e1332bf