Allowing refinement of final (or frozen) attributes of final local variables
Summary:
Refusing to refine attributes of local variables has increased the soundness of our type system at the cost of significantly increasing the friction of common operations.  Unfortunately, we also didn't have a great story for how to fix these issues, other than assigning to an intermediate variable, which isn't always nice.

Luckily enough for us there are already features in the runtime and our type system that can allow for sound refinement of attributes: `Final` and frozen dataclasses.

The final or frozen attribute of a local final variable are effectively local finals themselves, and therefore can be safely refined.
We just must ensure that an attribute comes from a line of all finals.

It's kind of unfortunate we can't support
```
def foo() -> None:
   x = makes_x()
   if x.final_attr is not None:
       expects_non_optional(x)
```
without explicitly marking x as Final, but without that mark, we'd need to build out facilities to invalidate these attribute refinements when the base object is reassigned, which is out of scope of this diff.

This approach relies on driving adoption of Final, which is good anyway.

Reviewed By: dkgi

Differential Revision: D18278440

fbshipit-source-id: 672e69104a1dde6d65a15a1ad819a8aaec0ac2cf